+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.4.0] - 2022-02-01
+### Added:
+- getTimestamp method DateTimeHelper
+- subtractDateInterval method DateTimeHelper
+
+## [1.3.0]
 ### Changed
 - Moved const declaration above properties 
 
@@ -25,7 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Project init
 
 
-[Unreleased]: https://github.com/studioraz/magento2-base/compare/1.2.0...HEAD
-[1.2.0]: https://github.com/studioraz/magento2-base/compare/1.1.0...1.2.0
-[1.1.0]: https://github.com/studioraz/magento2-base/compare/1.0.0...1.1.0
+[Unreleased]: https://github.com/studioraz/magento2-base/compare/1.4.0...HEAD
+[1.4.0]: https://github.com/studioraz/magento2-base/releases/tag/1.4.0
+[1.3.0]: https://github.com/studioraz/magento2-base/releases/tag/1.3.0
+[1.2.0]: https://github.com/studioraz/magento2-base/releases/tag/1.2.0
+[1.1.0]: https://github.com/studioraz/magento2-base/releases/tag/1.1.0
 [1.0.0]: https://github.com/studioraz/magento2-base/releases/tag/1.0.0